add: registry as separate plugin
