+++ src/kubedash/lib_plugins/registry.py
@@ -1,6 +1,9 @@
 #!/usr/bin/env python3
 
-import  requests, json, hashlib
+from flask import Blueprint, request, session,render_template, redirect, url_for, flash, jsonify
+from flask_login import login_required
+
+import  requests, json, hashlib, datetime
 from itsdangerous import base64_encode, base64_decode
 from lib_functions.components import db
 from lib_functions.helper_functions import get_logger, ErrorHandler, ResponseHandler
@@ -8,12 +11,18 @@ from flask_login import UserMixin
 from sqlalchemy import inspect
 from datetime import datetime
 
+from lib_functions.helper_functions import get_logger, ErrorHandler
+from lib_functions.components import csrf
+
 ##############################################################
-## Helper Functions
+## variables
 ##############################################################
 
+registry = Blueprint("registry", __name__)
 logger = get_logger()
 
+#############################################################
+## Helper Functions
 ##############################################################
 
 def get_base_url(registry_server_url: str) -> str:
@@ -145,10 +154,10 @@ def get_image_sbom_vulns(registry_server_url, image, tag) -> list:
                 vulnerabilities.append(vulnerability)
         return vulnerabilities
 
+##############################################################
+# OCI Registry Functions
+##############################################################
 
-##############################################################
-## Registry Data Functions
-##############################################################
 def RegistryGetRepositories(registry_server_url: str) -> list:
     """Get all repositories from the registry server
     
@@ -634,3 +643,143 @@ def RegistryGetEvent(repository, tag):
     if inspector.has_table("registry_events"):
         registry_events = RegistryEvents.query.filter_by(repository=repository, tag=tag).all()
     return registry_events
+
+##############################################################
+# OCI Registry Routes
+##############################################################
+
+
+@registry.route("/registry", methods=['GET', 'POST'])
+@login_required
+def registry_main():
+    selected = None
+    registry_server_auth_user = None
+    registry_server_auth_pass = None
+    registry_server_auth = False
+    if request.method == 'POST':
+        selected = request.form.get('selected')
+        request_type = request.form['request_type']
+        if request_type == "create":
+            registry_server_tls = request.form.get('registry_server_tls_register_value') in ['True']
+            insecure_tls = request.form.get('insecure_tls_register_value') in ['True']
+            registry_server_url = request.form.get('registry_server_url')
+            registry_server_port = request.form.get('registry_server_port')
+            if request.form.get('registry_server_auth_user') and request.form.get('registry_server_auth_pass'):
+                registry_server_auth_user = request.form.get('registry_server_auth_user')
+                registry_server_auth_pass = request.form.get('registry_server_auth_pass') # bas64 encoded
+                registry_server_auth = True
+
+            RegistryServerCreate(registry_server_url, registry_server_port, registry_server_auth, 
+                        registry_server_tls, insecure_tls, registry_server_auth_user, 
+                        registry_server_auth_pass)
+            flash("Registry Created Successfully", "success")
+        elif request_type == "edit":
+            registry_server_tls = request.form.get('registry_server_tls_edit_value') in ['True']
+            insecure_tls = request.form.get('insecure_tls_edit_value') in ['True']
+            registry_server_url = request.form.get('registry_server_url')
+            registry_server_url_old = request.form.get('registry_server_url_old')
+            registry_server_port = request.form.get('registry_server_port')
+            registry_server_auth = request.form.get('registry_server_auth')
+            if request.form.get('registry_server_auth_user') and request.form.get('registry_server_auth_pass'):
+                registry_server_auth_user = request.form.get('registry_server_auth_user')
+                registry_server_auth_pass = request.form.get('registry_server_auth_pass')
+                registry_server_auth = True
+
+            RegistryServerUpdate(registry_server_url, registry_server_url_old, registry_server_port, 
+                        registry_server_auth, registry_server_tls, insecure_tls, registry_server_auth_user, 
+                        registry_server_auth_pass)
+            flash("Registry Updated Successfully", "success")
+        elif request_type == "delete":
+            registry_server_url = request.form.get('registry_server_url')
+
+            RegistryServerDelete(registry_server_url)
+            flash("Registry Deleted Successfully", "success")
+
+    registries = RegistryServerListGet()
+
+    return render_template(
+      'registry.html.j2',
+        registries = registries,
+        selected = selected,
+    )
+
+@registry.route("/image/list", methods=['GET', 'POST'])
+@login_required
+def image_list():
+    selected = None
+    if request.method == 'POST':
+        selected = request.form.get('selected')
+        session['registry_server_url'] = request.form.get('registry_server_url')
+
+    image_list = RegistryGetRepositories(session['registry_server_url'])
+
+    return render_template(
+        'registry-image-list.html.j2',
+        image_list = image_list,
+        selected = selected,
+    )
+    
+@registry.route("/image/tags", methods=['GET', 'POST'])
+@login_required
+def image_tags():
+    selected = None
+    if request.method == 'POST':
+        selected = request.form.get('selected')
+        if 'image_name' in request.form:
+            session['image_name'] = request.form.get('image_name')
+
+    tag_list = RegistryGetTags(session['registry_server_url'], session['image_name'])
+
+    return render_template(
+        'registry-image-tag-list.html.j2',
+        tag_list = tag_list,
+        selected = selected,
+        image_name = session['image_name'],
+    )
+
+@registry.route("/image/tag/delete", methods=['GET', 'POST'])
+@login_required
+def image_tag_delete():
+    if request.method == 'POST':
+        tag_name = request.form.get('tag_name')
+        image_name = request.form.get('image_name')
+        RegistryDeleteTag(session['registry_server_url'], image_name, tag_name)
+        return redirect(url_for('registry.image_tags'), code=307)
+    else:
+        return redirect(url_for('registry.login'))
+
+@registry.route("/image/data", methods=['GET', 'POST'])
+@login_required
+def image_data():
+    if request.method == 'POST':
+        if 'tag_name' in request.form:
+            session['tag_name'] = request.form.get('tag_name')
+
+    tag_data = RegistryGetManifest(session['registry_server_url'], session['image_name'], session['tag_name'])
+    tag_events = RegistryGetEvent(session['image_name'], session['tag_name'])
+
+    return render_template(
+        'registry-image-tag-data.html.j2',
+        tag_data = tag_data,
+        tag_events = tag_events,
+        image_name = session['image_name'],
+        tag_name = session['tag_name'],
+    )
+
+@registry.route("/registry/events", methods=['POST'])
+@csrf.exempt
+def registry_events():
+    events = request.json["events"]
+    for event in events:
+        timestamp = datetime.now()
+        try:
+            actor = event["actor"]["name"]
+        except KeyError:
+            actor = None
+        if "tag" in event["target"]:
+            if event["request"]["useragent"] != "KubeDash":
+                RegistryEventCreate(event["action"], event["target"]["repository"], 
+                event["target"]["tag"], event["target"]["digest"], event["request"]["addr"].split(":")[0], actor, timestamp)
+
+    resp = jsonify(success=True)
+    return resp
